Revert "Integrate frontend and backend"
