+++ backend/package.json
@@ -1,18 +1,15 @@
-{
-  "name": "cs410-final-project",
-  "version": "1.0.0",
-  "description": "Team: literallyAnything",
-  "main": "index.js",
-  "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
-  },
-  "keywords": [],
-  "author": "",
-  "license": "ISC",
-  "dependencies": {
-    "axios": "^1.6.2",
-    "cors": "^2.8.5",
-    "ejs": "^3.1.9",
-    "express": "^4.18.2"
-  }
-}
+{"name": "cs410-final-project",
+"version": "1.0.0",
+"description": "Team: literallyAnything",
+"main": "index.js",
+"scripts": {
+  "test": "echo \"Error: no test specified\" && exit 1"
+},
+"keywords": [],
+"author": "",
+"license": "ISC",
+"dependencies": {
+  "axios": "^1.6.2",
+  "cors": "^2.8.5",
+  "ejs": "^3.1.9",
+  "express": "^4.18.2"}}
