+++ package.json
@@ -1,4 +1,21 @@
 {
+<<<<<<< HEAD
+  "name": "cs410-final-project",
+  "version": "1.0.0",
+  "description": "Team: literallyAnything",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "axios": "^1.6.2",
+    "cors": "^2.8.5",
+    "ejs": "^3.1.9",
+    "express": "^4.18.2"
+=======
   "name": "react-chrome-ext",
   "version": "0.1.0",
   "private": true,
@@ -11,7 +28,6 @@
     "@types/node": "^16.18.11",
     "@types/react": "^18.0.27",
     "@types/react-dom": "^18.0.10",
-    "axios": "^1.6.2",
     "copy-webpack-plugin": "^11.0.0",
     "css-loader": "^6.7.3",
     "html-webpack-plugin": "^5.5.0",
@@ -29,5 +45,6 @@
     "ts-loader": "^9.4.2",
     "webpack": "^5.75.0",
     "webpack-cli": "^5.0.1"
+>>>>>>> b901fa953ab11c8168637cc852bd72584c07cdfb
   }
 }
