--- views/index.ejs
@@ -1,24 +1,0 @@
-<!DOCTYPE html>
-<html lang="en">
-<head>
-    <meta charset="UTF-8">
-    <meta name="viewport" content="width=device-width, initial-scale=1.0">
-    <title>Chrome Extension Backend</title>
-</head>
-<body>
-    <h1>Chrome Extension Backend</h1>
-    
-    <form action="/get-remote-text" method="post">
-        <label for="url">Enter URL:</label>
-        <input type="text" id="url" name="url" required>
-        <button type="submit">Submit</button>
-    </form>
-
-    <div id="result"></div>
-    
-
-    <script>
-        // Add any client-side scripting if needed
-    </script>
-</body>
-</html>
